smtp-server: push the metadata to lib.rs
