--- smtp-server/src/metadata.rs
@@ -1,11 +1,0 @@
-use smtp_message::Email;
-
-pub struct MailMetadata<U> {
-    pub user: U,
-    pub from: Option<Email>,
-    pub to: Vec<Email>,
-}
-
-pub struct ConnectionMetadata<U> {
-    pub user: U,
-}
